+++ src/game/level-creation/generateRandomLevel.ts
@@ -19,6 +19,7 @@ export type LevelSettings = {
   extraPlacementLimits?: number;
   buffers?: number;
   bufferSizes?: number;
+  blockColorPick?: "start" | "end";
   bufferPlacementLimits?: number;
   extraBuffers?: { size: number; amount: number; limit: number }[];
   playMoves?: [minCount: number, maxPercent: number];
@@ -34,6 +35,7 @@ export const generateRandomLevel = (
     bufferSizes = 1,
     bufferPlacementLimits = 0,
     extraBuffers = [],
+    blockColorPick = "start",
     hideBlockTypes = "none",
     stacksPerColor = 1,
   }: LevelSettings,
@@ -42,7 +44,10 @@ export const generateRandomLevel = (
   // This results in gradually reveal new colors as you progress
   const colorPool = Math.ceil(amountColors / 4) * 4;
   // Generate level, this should be extracted
-  const availableColors = BLOCK_COLORS.slice(0, colorPool);
+  const availableColors =
+    blockColorPick === "end"
+      ? BLOCK_COLORS.slice(-colorPool)
+      : BLOCK_COLORS.slice(0, colorPool);
   shuffle(availableColors, random);
 
   // This simulates having 12 colors for the first few levels,