+++ src/ui/Background/Background.stories.tsx
@@ -17,6 +17,10 @@ const meta = {
       options: ["none", "easy", "hard", "special", "scrambled"],
       control: "select",
     },
+    theme: {
+      options: ["default", "halloween", "winter"],
+      control: "select",
+    },
   },
   // More on argTypes: https://storybook.js.org/docs/api/argtypes
   // Use `fn` to spy on the onClick arg, which will appear in the actions panel once invoked: https://storybook.js.org/docs/essentials/actions#action-args