+++ src/ui/Background/Background.tsx
@@ -52,12 +52,60 @@ const snow: ISourceOptions = {
   },
 };
 
+const ghosts: ISourceOptions = {
+  detectRetina: true,
+  fpsLimit: 60,
+  particles: {
+    number: {
+      value: 10,
+    },
+    move: {
+      direction: MoveDirection.top,
+      enable: true,
+      random: false,
+      straight: false,
+    },
+    opacity: {
+      value: { min: 0.1, max: 0.5 },
+    },
+    size: {
+      value: { min: 3, max: 20 },
+    },
+    shape: {
+      type: "emoji",
+      options: {
+        emoji: {
+          font: "system-ui",
+          value: "️👻",
+        },
+      },
+    },
+
+    wobble: {
+      distance: 20,
+      enable: true,
+      speed: {
+        min: -5,
+        max: 5,
+      },
+    },
+  },
+};
+
+const themeParticles: Record<BlockTheme, ISourceOptions | undefined> = {
+  winter: snow,
+  halloween: ghosts,
+  default: undefined,
+};
+
 export const Background: React.FC<PropsWithChildren<Props>> = ({
   children,
   levelType,
+  theme = "default",
 }) => {
   const [init, setInit] = useState(false);
-  const [useParticles] = useState(false);
+
+  const particles = themeParticles[theme];
 
   // this should be run only once per application lifetime
   useEffect(() => {
@@ -81,9 +129,23 @@ export const Background: React.FC<PropsWithChildren<Props>> = ({
         }
       )}
     >
+      {theme === "halloween" && (
+        <div className="absolute left-0 top-0 h-full w-full animate-lightning bg-white"></div>
+      )}
       <div className="absolute left-0 top-0 h-full w-full mix-blend-multiply bg-wood-texture"></div>
+      {theme === "halloween" && (
+        <div className="absolute left-0 bottom-0 text-8xl -rotate-12">🎃</div>
+      )}
+      {theme === "winter" && (
+        <div className="absolute left-0 bottom-0 text-8xl">🎄</div>
+      )}
+      {theme === "winter" && (
+        <div className="absolute right-0 top-3/4 text-8xl -rotate-45 translate-x-8">
+          ⛄️
+        </div>
+      )}
       <div className="absolute left-0 top-0 h-safe-area w-full">{children}</div>
-      {init && useParticles && <Particles id="tsparticles" options={snow} />}
+      {init && particles && <Particles id="tsparticles" options={particles} />}
     </div>
   );
 };